Modify timebucketizer to include record retrieval and fixed pr changes
