+++ hasher-matcher-actioner/setup.py
@@ -12,7 +12,7 @@ extras_require = {
 }
 
 all_extras = set(sum(extras_require.values(), []))
-extras_require["test"] = sorted({"pytest==6.2.1"} | all_extras)
+extras_require["test"] = sorted({"pytest==6.2.1", "freezegun==1.1.0"} | all_extras)
 extras_require["package"] = ["wheel"]
 extras_require["lint"] = ["black==21.12b0"]
 extras_require["type"] = ["types-requests==2.27.1"]
